% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annotations}
\alias{annotations}
\title{A list of annotations with which to intersect user-provided SNPs}
\format{
A list of BED tibbles containing various functional annotations of the genome
}
\usage{
annotations
}
\description{
A list of annotations with which to intersect user-provided SNPs
}
\section{annotations[['TFBSs']]}{


A valr-compatible tibble BED object imported from BED files of the transcription factor binding sites from different ReMap 2020 experiments
(Internal data example)

Source: /working/lab_georgiat/alexandT/target_gene_prediction_paper/output/ReMap/split_by_Experiment.TF.Biotype/

\describe{
  \item{chrom}{chromosome}
  \item{start}{start position}
  \item{end}{end position}
  \item{Experiment}{Accession number of experiment from which the data was generated (e.g. GENCODE)}
  \item{TranscriptionFactor}{Transcription factor assayed}
  \item{CellType}{Cell type assayed}
}
}

\section{annotations[['DHSs']]}{


A valr-compatible tibble BED object imported from BED files of the DHSs binned by H3K27ac/H3K4me1 mark signal/specificity for each cell type.
(Internal data example)

Source: /working/lab_georgiat/jonathB/PROJECTS/trench_lab/target_gene_prediction/output/activity_signal_matrix/bin_regions/

\describe{
  \item{chrom}{chromosome}
  \item{start}{start position}
  \item{end}{end position}
  \item{CellType}{Cell type assayed}
  \item{Mark}{Histone modification mark assayed}
  \item{Bin}{What bin does this DHS rank in, according to the signal/specificity level of its marks?}
  \item{Binning}{Number of bins into which data is split (e.g. quartiles, deciles)}
  \item{Method}{Signal or specificity ranking of DHS sites}
}
}

\keyword{datasets}
