% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BCVars}
\alias{BCVars}
\title{Fine-mapped breast cancer SNPs from BCAC (user-input data example)}
\format{
A GRanges object with 5369 ranges and 4 metadata columns:
\describe{
  \item{variant}{Unique SNP ID (rsID / position and allele information)}
  \item{cs}{Credible set that the SNP belongs to}
  \item{trait}{Trait associated with the SNP from GWAS (breast cancer)}
  \item{status}{Specific breast cancer (estrogen receptor) phenotype associated with the SNP}
}
}
\source{
{/working/lab_georgiat/alexandT/target_gene_prediction_paper/output/Traits/BC/BC.VariantList.bed}
}
\usage{
BCVars
}
\description{
A GRanges object from a BED file of the strong fine-mapped breast cancer SNPs within credible set at breast cancer signals.
This is a template for the user-inputted trait variants file.
}
\keyword{datasets}
