% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_fisher_grouped.R
\name{bed_fisher_grouped}
\alias{bed_fisher_grouped}
\title{Fisher test of significance of overlaps between grouped sets of intervals}
\usage{
bed_fisher_grouped(bedA, bedB, genome, bedA_groups = NULL)
}
\arguments{
\item{bedA}{A tibble BED with columns: chrom, start, end, ...(metadata cols)...}

\item{bedB}{A tibble BED with columns: chrom, start, end, ...(metadata cols)...}

\item{genome}{A tibble BED of chromosome sizes with columns: chrom, size}

\item{bedA_groups}{A character vector of data variable(s) in bedA upon which to group data and perform the test on each group of intervals separately}
}
\value{
The outcome of the Fisher test(s) on the overlaps between bedA and bedB
}
\description{
Calculate Fisher's test on the number of intervals that are shared and unique between two input tibble BEDs.
If a grouping variable is provided for bedA, Fisher tests can be performed within these groups of intervals in bedA.
}
\details{
This function is built as a modification of the valr::bed_fisher() function.
Equivalent to bedtools fisher -a A.bed -b B.bed (performed within each variable group for A.bed)
}
