% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_intersect_left.R
\name{bed_intersect_left}
\alias{bed_intersect_left}
\title{Left join intersection of two tibble BEDs}
\usage{
bed_intersect_left(bedA, bedB, suffix = c("", ""))
}
\arguments{
\item{bedA}{A tibble BED with columns: chrom, start, end, ...(metadata cols)...}

\item{bedB}{A tibble BED with columns: chrom, start, end, ...(metadata cols)...}

\item{suffix}{A character vector of suffixes to add to metadata column names. (If the two input tibble BEDs share any common metadata column names, suffixes must be provided.)}
}
\value{
A tibble BED of all intervals from bedA that contain intersects with bedB, plus all metadata columns from both
}
\description{
Intersect two tibble BEDs and retain the full intervals from the lefthand tibble BED that have any overlap.
If the two input tibble BEDs share any common metadata column names, suffixes must be provided.
This function is built as a modification of the valr::bed_intersect() function.
Equivalent to bedtools -a A -b B -wa.
}
