% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_and_weight_and_widen_annotations.R
\name{bind_and_weight_and_widen_annotations}
\alias{bind_and_weight_and_widen_annotations}
\title{Bind and weight and widen annotation tibbles}
\usage{
bind_and_weight_and_widen_annotations(id_cols, annotation_level, ...)
}
\arguments{
\item{id_cols}{Vector of column names that uniquely identifies each observation, sent to pivot_wider(id_cols)}

\item{...}{The long-format annotation tibble(s) to bind and widen}

\item{annotation.level}{Name of the annotation level (e.g. pair, gene, variant) that the annotations fall within}
}
\value{
Wide-format annotation dataframe, with only one observation per row and only one annotation per column
}
\description{
Bind annotation tibbles, optionally weight values by a weighting column, and then apply pivot_wider
to the annotation columns, converting from long-format to wide-format, with only 1 variant or gene or
variant-gene pair per row and 1 annotation per column
}
