% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bins}
\alias{bins}
\title{DHSs binned by H3K27ac/H3K4me1 mark signal/specificity (internal data example)}
\format{
A tibble BED object with 5369 ranges and 4 metadata columns:
\describe{
  \item{chrom}{chromosome}
  \item{start}{start position}
  \item{end}{end position}
  \item{CellType}{Cell type assayed}
  \item{Mark}{Histone modification mark assayed}
  \item{Bin}{What bin does this DHS rank in, according to the signal/specificity level of its marks?}
  \item{Binning}{Number of bins into which data is split (e.g. quartiles, deciles)}
  \item{Method}{Signal or specificity ranking of DHS sites}
}
}
\source{
{/working/lab_georgiat/jonathB/PROJECTS/trench_lab/target_gene_prediction/output/activity_signal_matrix/bin_regions/bin_regions_by_specificity/sorted/H3K27ac/deciles/}
}
\usage{
bins
}
\description{
A valr-compatible tibble BED object imported from BED files of the DHSs binned by H3K27ac/H3K4me1 mark signal/specificity for each cell type
Data only given for MCF-7 currently
}
\keyword{datasets}
