% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_driver_symbols.R
\name{check_driver_symbols}
\alias{check_driver_symbols}
\title{check that all symbols are in the GENCODE data}
\usage{
check_driver_symbols(drivers, driversFile)
}
\arguments{
\item{drivers}{vector of driver symbols}

\item{driversFile}{name of the drivers file}
}
\value{
A file of variant-gene pair predictions, with associated scores, saved in the given output directory.
}
\description{
check that all symbols are in the GENCODE data
}
