% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ChrSizes}
\alias{ChrSizes}
\title{A valr-compatible genome dataframe of chromosome sizes from hg19}
\format{
Tibble
\describe{
  \item{chrom}{chromosome}
  \item{size}{chromosome length}
}
}
\source{
{/working/lab_georgiat/alexandT/target_gene_prediction_paper/data/reference_data/hg19.genome}
}
\usage{
ChrSizes
}
\description{
A valr-compatible genome dataframe of chromosome sizes from hg19
}
\keyword{datasets}
