% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contact_metadata}
\alias{contact_metadata}
\title{Metadata for the contact data samples that accompany the package}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 27 rows and 5 columns.
}
\usage{
contact_metadata
}
\description{
Metadata matching each contact assay in the contact list to its study, celltype and tissue
}
\details{
\describe{
  \item{code}{Shortened cell type code in `DHSs` CellType column}
  \item{name}{Full, descriptive name of the cell type}
  \item{tissue}{Tissue of origin of the sample}
  \item{stage}{Developmental stage of the sample (embryo / fetus / adult)}
}
}
\keyword{datasets}
