% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DHSs}
\alias{DHSs}
\title{Mark specificity-/signal-binned DHSs across cell types}
\format{
An object of class \code{list} of length 2.
}
\source{
{/working/lab_georgiat/jonathB/PROJECTS/trench_lab/target_gene_prediction/output/activity_signal_matrix/bin_regions/}
}
\usage{
DHSs
}
\description{
A list of valr-compatible tibble BED objects imported from BED files of the DHSs binned by H3K27ac/H3K4me1
mark signal/specificity for each cell type.
(Internal data example)
}
\details{
\describe{
  \item{chrom}{chromosome}
  \item{start}{start position}
  \item{end}{end position}
  \item{CellType}{Cell type assayed}
  \item{Mark}{Histone modification mark assayed}
  \item{Bin}{What bin does this DHS rank in, according to the signal/specificity level of its marks?}
  \item{Binning}{Number of bins into which data is split (e.g. quartiles, deciles)}
  \item{Method}{Signal or specificity ranking of DHS sites}
}
}
\keyword{datasets}
