% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DHSs_metadata}
\alias{DHSs_metadata}
\title{Metadata for `DHSs`}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20 rows and 4 columns.
}
\usage{
DHSs_metadata
}
\description{
Metadata matching each DHS celltype code to its name, tissue and developmental stage
}
\details{
\describe{
  \item{code}{Shortened cell type code in `DHSs` CellType column}
  \item{name}{Full, descriptive name of the cell type}
  \item{tissue}{Tissue of origin of the sample}
  \item{stage}{Developmental stage of the sample (embryo / fetus / adult)}
}
}
\keyword{datasets}
