% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drivers}
\alias{drivers}
\title{Data frame of list(s) of driver gene symbols, which are treated as the positive control causal gene sets}
\format{
Tibble
\describe{
  \item{DriversList}{Name of the drivers list}
  \item{symbol}{Gene symbol of causal genes}
}
}
\source{
{/working/lab_georgiat/alexandT/target_gene_prediction_paper/data/Traits/BC/KnownGenes/}
}
\usage{
drivers
}
\description{
Data frame of list(s) of driver gene symbols, which are treated as the positive control causal gene sets
}
\keyword{datasets}
