% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PR.R
\name{get_PR}
\alias{get_PR}
\title{Get PR statistics for the predictions}
\usage{
get_PR(scores, txv_master, ...)
}
\arguments{
\item{...}{Columns to use as predictors}

\item{predictions}{`predictions` df in predict_target_genes function}
}
\value{
`PR` tibble
}
\description{
Get PR and AUPRC for predictions
}
