% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hichip}
\alias{hichip}
\title{Paired-end BEDPE list object}
\format{
A list of paired BED-format tibbles with 359,970 interaction pairs (719,940 intervals) and 2 metadata columns:
\describe{
  \item{chrom}{chromosome}
  \item{start}{start position}
  \item{end}{end position}
  \item{InteractionID}{ID that links the paired ends between elements (first-last)}
  \item{OldPairID}{Old ID}
  \item{value}{Reads}
}
}
\source{
{/working/lab_georgiat/alexandT/target_gene_prediction/data/HiChIP/MCF7/abc-ready/HiChIP.bedpe}
}
\usage{
hichip
}
\description{
A paired c(dataframe of first end, dataframe of last end) list object representation of the MCF-7 HiChIP BEDPE file with an ID column matching pairs and metadata columns
Data only given for MCF-7 currently
}
\keyword{datasets}
