% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_BED_to_GRanges.R
\name{import_BED_to_GRanges}
\alias{import_BED_to_GRanges}
\title{Import a BED}
\usage{
import_BED_to_GRanges(bedfile, metadata_cols = NULL)
}
\arguments{
\item{bedfile}{A tab-delimited file in bed format with columns: chr, start, stop, ...(metadata cols)...}

\item{metadata_cols}{A character vector of the names of metadata columns (bedfile cols 4+) in the BED file, to be imported as metadata in the GRanges object}
}
\value{
A GRanges object representation of the BED file, optionally with metadata columns
}
\description{
Import a BED file as a GRanges object
}
