% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_BEDPE_to_List.R
\name{import_BEDPE_to_List}
\alias{import_BEDPE_to_List}
\title{Import a BEDPE}
\usage{
import_BEDPE_to_List(bedpefile, metadata_cols = NULL)
}
\arguments{
\item{metadata_cols}{A character vector of the names of metadata columns (bedfile cols 4+) in the BED file}

\item{bedfile}{A tab-delimited file in bedpe format with columns: chrA, startA, stopA, chrB, startB, stopB, ...(metadata cols)...}
}
\value{
A paired (first + last) list object representation of the BEDPE file with an ID column matching pairs, optionally with metadata columns
}
\description{
Import a paired-end BEDPE file as a list object c(dataframe of first end, dataframe of last end)
}
