% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_variants_GRanges.R
\name{import_variants_GRanges}
\alias{import_variants_GRanges}
\title{Import variants BED}
\usage{
import_variants_GRanges(x)
}
\arguments{
\item{x}{A tab-delimited file in bed format with columns: chr, start, stop, snpid, locus}
}
\value{
A GRanges object representation of the variant BED file
}
\description{
Import a BED file of trait-associated variants grouped by association signal, for example SNPs correlated with an index variant, or credible sets of fine-mapped variants
}
