% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_annotations.R
\name{intersect_annotations}
\alias{intersect_annotations}
\title{Intersect with the DHS annotations}
\usage{
intersect_annotations(query, annots = annotations)
}
\arguments{
\item{query}{A query bed}

\item{annotations}{A list of BED tibbles to be intersected by the query bed (metadata columns will form the annotation name)}
}
\value{
A tibble of intersected columns, with .query and .annotation suffixes
}
\description{
Intersect a query BED with the DHS annotations.
}
