% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_annotations_list.R
\name{intersect_annotations_list}
\alias{intersect_annotations_list}
\title{Intersect with the annotations list}
\usage{
intersect_annotations_list(query)
}
\arguments{
\item{query}{A query bed}
}
\value{
A tibble of intersected columns, with .query and .annotation suffixes
}
\description{
Intersect a query BED with the annotations list.
}
