% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_DHSs.R
\name{intersect_DHSs}
\alias{intersect_DHSs}
\title{Intersect with the DHS annotations}
\usage{
intersect_DHSs(l, query, DHSs, ...)
}
\arguments{
\item{query}{A query bed}

\item{DHSs}{A BED tibble (the DHSs) to be intersected by the query bed}

\item{...}{query columns to be retained in the output}
}
\value{
A tibble of intersected columns, with .query and .annotation suffixes
}
\description{
Intersect a query BED with the DHS annotations.
}
