% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_contact_data.R
\name{load_contact_data}
\alias{load_contact_data}
\title{Load contact (3C) data}
\usage{
load_contact_data(contactDir)
}
\arguments{
\item{contactDir}{The directory in which the package's accompanying contact data is stored.
Files must have columns ( chrA | startA | endA | chrB | startB | endB | score | celltype ) and *.bedpe file extension.}
}
\value{
`contact` object (list of BEDPE lists containing loops and their normalised scores for multiple cell types)
}
\description{
Load the directory of contact data, downloaded from an FTP(?) alongside the package.
The function will take all files in this directory ending in *.bedpe, convert them to
BEDPE list objects, add an InteractionID column and normalise the score column within
each sample.
}
