% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_DHSs_data.R
\name{load_DHSs_data}
\alias{load_DHSs_data}
\title{Load DHSs data}
\usage{
load_DHSs_data(referenceDir)
}
\arguments{
\item{referenceDir}{The directory in which the package's accompanying data is stored.
(DHSs data is in the DHSs/ subdirectory).
Files must have columns ( chr | start | end | decile ) and *.bed file extension.}
}
\value{
`DHSs` object (list of BED tibbles containing binned DHSs)
}
\description{
Load the directory of DHSs data, downloaded from an FTP(?) alongside the package.
The function will take all files in this directory ending in *.bed, convert them to
BED tibble objects and add metadata.
}
