% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_variants.R
\name{overlap_variants}
\alias{overlap_variants}
\title{Overlap variants with another BED file}
\usage{
overlap_variants(vars, bed_to_overlap)
}
\arguments{
\item{vars}{A GRanges object of trait variants with metadata columns snpid and locus}

\item{bed_to_overlap}{A GRanges object of genomic annotations}
}
\value{
A GRanges object representation of the variant BED file
}
\description{
Function to overlap the user-provided trait variants file with other relevant BED files from package data
}
