% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursively_bind_rows.R
\name{recursively_bind_rows}
\alias{recursively_bind_rows}
\title{Recursively bind rows}
\usage{
recursively_bind_rows(nested_list, nest_names = NULL)
}
\arguments{
\item{nested_list}{The nested list}

\item{nest_names}{The names of each successive nesting of the list. If none provided, nest1, nest2, nest3, ... will be used}
}
\value{
A recursively row-bound tibble, with nested list names as added ID columns.
}
\description{
Recursively bind rows of a nested list.
}
