% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{remap}
\alias{remap}
\title{ReMap TFBSs (internal data example)}
\format{
A GRanges object with 118576 ranges and 3 metadata columns:
\describe{
  \item{seqnames}{chromosome}
  \item{ranges}{start-end}
  \item{strand}{*/+/-}
  \item{Experiment}{Accession number of experiment from which the data was generated (e.g. GENCODE)}
  \item{TranscriptionFactor}{Transcription factor assayed}
  \item{CellType}{Cell type assayed}
}
}
\source{
{/working/lab_georgiat/alexandT/target_gene_prediction_paper/output/ReMap/split_by_Experiment.TF.Biotype/}
}
\usage{
remap
}
\description{
A GRanges object imported from BED files of the transcription factor binding sites from different ReMap 2020 experiments
}
\keyword{datasets}
