% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TSSs}
\alias{TSSs}
\title{A tibble BED object of transcription start sites of all transcripts (from GENCODE)}
\format{
Tibble
\describe{
  \item{chrom}{chromosome}
  \item{start}{start position}
  \item{end}{end position}
  \item{ensg}{ENSEMBL gene ID}
  \item{symbol}{Gene symbol of causal genes}
  \item{enst}{ENSEMBL transcript ID}
}
}
\source{
{/working/lab_georgiat/alexandT/target_gene_prediction_paper/data/reference_data/GENCODE/gencode.v34lift37.basic.tss.bed.gz}
}
\usage{
TSSs
}
\description{
A tibble BED object of transcription start sites of all transcripts (from GENCODE)
}
\keyword{datasets}
