% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_performance.R
\name{get_performance}
\alias{get_performance}
\title{Get performance metrics for the predictions}
\usage{
get_performance(predictions, prediction_col)
}
\arguments{
\item{predictions}{Scores df in predict_target_genes function}
}
\value{
`performance` object (performance)
}
\description{
Limit predictions to credible sets with at least one driver within the maximum prediction distance.
Calculate p-value, OR, precision, recall, sensitivity, specificity and F-score for the prediction.
}
