% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PR.R
\name{get_PR}
\alias{get_PR}
\title{Get PR statistics for the predictions}
\usage{
get_PR(scores, txv_master, drivers)
}
\arguments{
\item{scores}{scores object passed from predict_target_genes}

\item{txv_master}{txv_master object passed from predict_target_genes}

\item{drivers}{drivers object passed from predict_target_genes}
}
\value{
`PR` tibble
}
\description{
Get PR and AUPRC for predictions
}
