% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hichip}
\alias{hichip}
\title{Paired-end BEDPE list object c(dataframe of first end, dataframe of last end)}
\format{
A list of paired BED-format dataframes with 359,970 pairs (719940, ends) and 2 metadata columns:
\describe{
  \item{chr}{chromosome}
  \item{start}{start position}
  \item{end}{end position}
  \item{group}{Group column}
  \item{value}{Reads}
  \item{PairID}{Numeric ID that links the paired ends in each element (first-last)}
}
}
\source{
{/working/lab_georgiat/alexandT/target_gene_prediction/data/HiChIP/MCF7/abc-ready/HiChIP.bedpe}
}
\usage{
hichip
}
\description{
A paired (first + last) list object representation of the MCF-7 HiChIP BEDPE file with an ID column matching pairs and metadata columns
Data only given for MCF-7 currently
This is not a GRanges object, as the GenomicRanges package does not appear to support paired-ends BED representation
}
\keyword{datasets}
