% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_target_genes.R
\name{predict_target_genes}
\alias{predict_target_genes}
\title{Predict target genes of a list of fine-mapped non-coding SNPs for a trait}
\usage{
predict_target_genes(snpfile, trait = NULL, outdir = ".")
}
\arguments{
\item{snpfile}{A header-less BED file of fine-mapped trait-relevant SNPs, with metadata columns for the variant's name and the credible set it belongs to (chrom, start, stop, variant, cs)}

\item{trait}{The name of the trait of interest}

\item{outdir}{The output directory in which to save the predictions. If none given, saved in current directory.}
}
\value{
A file of variant-gene pair predictions, with associated scores, saved in the given output directory.
}
\description{
The master, user-facing function of this package.
}
