% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_target_genes.R
\name{predict_target_genes}
\alias{predict_target_genes}
\title{Predict target genes of a list of fine-mapped non-coding variants for a trait}
\usage{
predict_target_genes(
  trait = NULL,
  tissue = NULL,
  outDir = "out",
 
    variantsFile = "/working/lab_georgiat/alexandT/target.gene.prediction.package/external_data/reference/BC.VariantList.bed",
 
    driversFile = "/working/lab_georgiat/alexandT/target.gene.prediction.package/external_data/reference/breast_cancer_drivers_2021.txt",
 
    referenceDir = "/working/lab_georgiat/alexandT/target.gene.prediction.package/external_data/reference/",
  variant_to_gene_max_distance = 2e+06,
  min_proportion_of_variants_in_top_DHSs = 0.05,
  do_all_cells = F,
  do_scoring = F,
  do_performance = F,
  do_XGBoost = F
)
}
\arguments{
\item{trait}{Optional. The name of the trait of interest.}

\item{tissue}{Optional. The tissue(s) of action for the trait.}

\item{outDir}{The output directory in which to save the predictions. Default is "./out".}

\item{variantsFile}{A BED file of trait-associated variants grouped by association signal, for example SNPs correlated with an index variant, or credible sets of fine-mapped variants}

\item{driversFile}{The file containing a list of trait driver gene symbols.}

\item{referenceDir}{The directory containing the external, accompanying reference data.}

\item{variant_to_gene_max_distance}{The maximum absolute distance (bp) across which variant-gene pairs are considered. Default is 2Mb. The contact data is also already filtered to 2Mb.}

\item{min_proportion_of_variants_in_top_DHSs}{A threshold propportion of variants that reside in the specific DHSs of a celltype for that celltype to be considered enriched. Default is 5% (0.05).}

\item{do_all_cells}{If TRUE, the package will consider annotations across all available cell types, not just those with enriched enhancer variants. Default is FALSE.}
}
\value{
A file of variant-gene pair predictions, with associated scores, saved in the given output directory.
}
\description{
The master, user-facing function of this package.
}
