% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_target_genes.R
\name{predict_target_genes}
\alias{predict_target_genes}
\title{Predict target genes of fine-mapped variants for a trait}
\usage{
predict_target_genes(
  trait = NULL,
  celltype_of_interest = NULL,
  tissue_of_interest = NULL,
  outDir = "out",
 
    variantsFile = "/working/lab_jonathb/alexandT/tgp/example_data/data/BC/BC.VariantList.bed",
 
    driversFile = "/working/lab_jonathb/alexandT/tgp/example_data/data/BC/breast_cancer_drivers_2021.txt",
  referenceDir = "/working/lab_jonathb/alexandT/tgp/reference_data/data/",
  variant_to_gene_max_distance = 2e+06,
  min_proportion_of_variants_in_top_H3K27ac = 0.05,
  do_all_celltypes = F,
  do_all_celltypes_in_enriched_tissue = T,
  do_manual_weighting = F,
  n_unique_manual_weights = NULL,
  do_scoring = T,
  do_performance = T,
  do_XGBoost = F,
  do_timestamp = F,
  contact = NULL,
  H3K27ac = NULL
)
}
\arguments{
\item{trait}{Optional. The name of the trait of interest.}

\item{celltype_of_interest}{Optional. The celltype(s) of interest for the trait. Only annotations in these celltypes will be used to make predictions. Argument(s) must match the names of celltypes in the metadata. Make sure the celltype of interest has coverage across all annotations (TADs, contact, expression, H3K27ac) in the metadata table.}

\item{tissue_of_interest}{Optional. The tissue(s) of interest for the trait. Only annotations in these tissues will be used to make predictions.  Argument(s) must match the names of tissues in the metadata.}

\item{outDir}{The output directory in which to save the predictions. Default is "./out".}

\item{variantsFile}{A BED file of trait-associated variants grouped by association signal, for example SNPs correlated with an index variant, or credible sets of fine-mapped variants}

\item{driversFile}{Optional. The file containing a list of trait driver gene symbols. If do_performance is TRUE, must provide a driversFile.}

\item{referenceDir}{The directory containing the external, accompanying reference data.}

\item{variant_to_gene_max_distance}{The maximum absolute distance (bp) across which variant-gene pairs are considered. Default is 2Mb. The contact data is also already filtered to 2Mb.}

\item{min_proportion_of_variants_in_top_H3K27ac}{A threshold proportion of variants that reside in the specific H3K27ac-x-DHSs of a celltype for that celltype to be considered enriched. Default is 5% (0.05).}

\item{do_all_celltypes}{If TRUE, the package will combine annotations across all available cell types, not just those with enriched enhancer variants. Default is FALSE.}

\item{do_all_celltypes_in_enriched_tissue}{If TRUE, the package will combine all annotations for the tissue of the enriched celltype(s), not just the specifically enriched celltype(s). Default is TRUE. Make sure the enriched celltype has coverage across all annotations (TADs, contact, expression, H3K27ac) in the metadata table.}

\item{do_manual_weighting}{If TRUE, runs the manual weighting chunk of the script (weight_and_score_manually) to test out different combinations of annotations to generate a score. Default is FALSE.}

\item{n_unique_manual_weights}{The number of unique weights for the do_manual_weighting chunk to consider. If NULL, the chunk will consider as many unique weights as there are to_add components. Default is NULL.}

\item{do_scoring}{If TRUE, runs the scoring chunk of the script, which combines all of the constituent MAE annotations into one score per transcript-variant pair. Default is FALSE.}

\item{do_performance}{If TRUE, runs the performance chunk of the script, which measures the performance of the score and each of its constituent annotations in predicting drivers as the targets of nearby variants. Default is FALSE.}

\item{do_XGBoost}{If TRUE, runs the XGBoost chunk of the script, which generates a model to predict the targets of variants from all available annotations and rates the importance of each annotation. Default is FALSE.}

\item{do_timestamp}{If TRUE, will save output into a subdirectory timestamped with the data/time of the run.}

\item{contact}{If you are repeatedly running predict_target_genes, you can load the `contact` object from the referenceDir into the global environment and pass it to the function to prevent redundant re-loading each call to predict_target_genes.}

\item{H3K27ac}{If you are repeatedly running predict_target_genes, you can load the `H3K27ac` object from the referenceDir into the global environment and pass it to the function to prevent redundant re-loading with each call to predict_target_genes.}
}
\value{
A MultiAssayExperiment object with one assay object per annotation, one row per variant-transcript pair and one column per cell type (or 'value' if it is a non-cell-type-specific annotation).
}
\description{
The master, user-facing function of this package.
}
